Make the /example page the new homepage!
